+++ unitas_ext/application_top.php
@@ -11,7 +11,7 @@
  */
 
 // Plugin constants
-define('PLUGIN_UNITAS_EXT_VERSION', '1.3.0');
+define('PLUGIN_UNITAS_EXT_VERSION', '1.4.0');
 define('PLUGIN_UNITAS_EXT_PATH', __DIR__);
 
 // Load installer